pyproject.toml
[tool.poetry]
name = "onion-demo"
version = "0.1.0"
description = ""
authors = ["sydneywu <sydneywhz@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
SQLAlchemy = "^2.0.28"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
databases = "^0.9.0"
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"
psycopg2 = "^2.9.9"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5.post1"
pytest-cov = "^4.1.0"
mypy = "^1.15.0"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"

[tool.poetry.scripts]
lint = "scripts.lint:run_mypy"
test = "pytest"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
